feat: using PORT1 variable refactor: moving console test to test_console
